worker/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
python-multipart
email-validator
requests
asyncpg
pydantic-sqlalchemy